+++ pkg/otel/common/otlp/scope.go
@@ -31,26 +31,6 @@ type ScopeIds struct {
 	DroppedAttributesCount int
 }
 
-// ToDo remove this function once metrics have been converted to the model v1
-func NewScopeIds(resSpansDT *arrow.StructType) (*ScopeIds, error) {
-	scopeID, scopeDT, err := arrowutils.StructFieldIDFromStruct(resSpansDT, constants.Scope)
-	if err != nil {
-		return nil, werror.Wrap(err)
-	}
-
-	nameID, _ := arrowutils.FieldIDFromStruct(scopeDT, constants.Name)
-	versionID, _ := arrowutils.FieldIDFromStruct(scopeDT, constants.Version)
-	droppedAttributesCountID, _ := arrowutils.FieldIDFromStruct(scopeDT, constants.DroppedAttributesCount)
-	attrsID, _ := arrowutils.FieldIDFromStruct(scopeDT, constants.ID)
-	return &ScopeIds{
-		Scope:                  scopeID,
-		Name:                   nameID,
-		Version:                versionID,
-		DroppedAttributesCount: droppedAttributesCountID,
-		ID:                     attrsID,
-	}, nil
-}
-
 func NewScopeIdsFromSchema(schema *arrow.Schema) (*ScopeIds, error) {
 	scopeID, scopeDT, err := arrowutils.StructFieldIDFromSchema(schema, constants.Scope)
 	if err != nil {
@@ -68,49 +48,6 @@ func NewScopeIdsFromSchema(schema *arrow.Schema) (*ScopeIds, error) {
 		DroppedAttributesCount: droppedAttributesCountID,
 		ID:                     ID,
 	}, nil
-}
-
-// ToDo remove this function once metrics have been converted to the model v1
-
-// UpdateScopeWith appends a scope into a given scope spans from an Arrow list of structs.
-func UpdateScopeWith(
-	s pcommon.InstrumentationScope,
-	listOfStructs *arrowutils.ListOfStructs,
-	row int,
-	ids *ScopeIds,
-	attrsStore *Attributes16Store,
-) error {
-	_, scopeArray, err := listOfStructs.StructByID(ids.Scope, row)
-	if err != nil {
-		return werror.WrapWithContext(err, map[string]interface{}{"row": row})
-	}
-	name, err := arrowutils.StringFromStruct(scopeArray, row, ids.Name)
-	if err != nil {
-		return werror.WrapWithContext(err, map[string]interface{}{"row": row})
-	}
-	version, err := arrowutils.StringFromStruct(scopeArray, row, ids.Version)
-	if err != nil {
-		return werror.WrapWithContext(err, map[string]interface{}{"row": row})
-	}
-	droppedAttributesCount, err := arrowutils.U32FromStruct(scopeArray, row, ids.DroppedAttributesCount)
-	if err != nil {
-		return werror.WrapWithContext(err, map[string]interface{}{"row": row})
-	}
-
-	attrsID, err := arrowutils.NullableU16FromStruct(scopeArray, row, ids.ID)
-	if err != nil {
-		return werror.WrapWithContext(err, map[string]interface{}{"row": row})
-	}
-	if attrsID != nil {
-		attrs := attrsStore.AttributesByDeltaID(*attrsID)
-		if attrs != nil {
-			attrs.CopyTo(s.Attributes())
-		}
-	}
-	s.SetName(name)
-	s.SetVersion(version)
-	s.SetDroppedAttributesCount(droppedAttributesCount)
-	return nil
 }
 
 func UpdateScopeFromRecord(
